Pythonize "rookify" application
