--- src/modules/migrate_monitors/main.py
@@ -1,4 +1,0 @@
-from modules.module import ModuleHandler
-
-class MigrateMonitorsHandler(ModuleHandler):
-    pass
