+++ src/rookify/modules/analyze_ceph/__init__.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from .main import AnalyzeCephHandler
 
 HANDLER_CLASS = AnalyzeCephHandler
