+++ src/rookify/modules/migrate_monitors/main.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from ..module import ModuleHandler
+
+class MigrateMonitorsHandler(ModuleHandler):
+    pass
