+++ src/rookify/modules/migrate_osds/__init__.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from .main import MigrateOSDsHandler
 
 HANDLER_CLASS = MigrateOSDsHandler
